creates RabbitMQClient & GoAriNatsClient interface Both are direct rips of http equivalents for the moment Same code was in ARIProxyManager before, this moves it, TestAMQP still needs refactoring -- is broken by this commit.
